dev: Added feature for adding & saving workouts on target device
