+++ app/src/main/java/com/github/ldcdorn/haelth/data/Exercise.kt
@@ -1,0 +1,12 @@
+package com.github.ldcdorn.haelth.data
+
+import java.sql.Date
+
+data class Exercise(
+    val name: String,
+    val type: String,
+    val weight: Int,
+    val amount: Int,
+    val date: Date
+)
+
